issue-665: move HumanSearchMatcher to Web.Extensions layer
`HumanSearchMatcher.DetermineMatch` computes view-model metadata
(badge field name + display snippet) and is called only from
`SearchResultMappingExtensions.ToHumanSearchViewModel`. Per design-rules
§1 (Application = business logic, not view-model helpers) and §14
(view-model mapping belongs in Web), move to `Humans.Web.Extensions`.

Addresses Claude review WARN on commit 57b73421.
